neurons_functions/requirements.txt
intel-numpy
redis
confluent-kafka
flask